requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pika==1.1.0
Pillow==6.2.1
psutil==5.6.3
Werkzeug==0.16.0
